Added a single sentence describing the purpose of each Basic plugin
Added a sentence to the beginning of each Basic plugin section
describing the general capabilities of each plugin.
Removed additional text discussing their utilizations in detail
which could be better suited to a tutorial.
Removed redundant details which appear in comments in the codebase
which will be autogenerated by Sphinx.